refactor: #684 auth_service を認証フロー境界で分割する
登録・セッション・プロフィール・退会をファイル分割し、
単一ファイルの巨大化を解消する。公開 API・DI は不変。
