+++ Backend/internal/services/auth_service.go
@@ -1,21 +1,9 @@
 package services
 
 import (
-	"Backend/domain/entity"
 	"Backend/domain/repository"
-	"Backend/internal/config"
-	"Backend/internal/middleware"
-	"Backend/internal/models"
-	"crypto/rand"
-	"encoding/base64"
-	"errors"
-	"fmt"
 	"log"
-	"os"
-	"strings"
-	"time"
 
-	"golang.org/x/crypto/bcrypt"
 	"gorm.io/gorm"
 )
 
@@ -82,102 +70,6 @@ func (s *AuthService) issueRefreshToken(userID uint) string {
 	return token
 }
 
-// DeleteAccount ユーザーアカウントとその全データを削除する（個人情報保護法第28条対応）
-func (s *AuthService) DeleteAccount(userID uint) error {
-	if s.deletion == nil {
-		s.rebuildDeletionService()
-	}
-	if s.deletion == nil {
-		return errors.New("database not configured")
-	}
-	// リフレッシュトークンの失効は DeleteUser のトランザクション内で行われる (#616)
-	return s.deletion.DeleteUser(userID, UserDeletionActor{Kind: "self"})
-}
-
-func collectUserSessionIDs(tx *gorm.DB, userID uint) ([]string, error) {
-	sessions := map[string]struct{}{}
-	collect := func(model any) error {
-		var ids []string
-		if err := tx.Model(model).
-			Where("user_id = ?", userID).
-			Distinct().
-			Pluck("session_id", &ids).Error; err != nil {
-			return err
-		}
-		for _, id := range ids {
-			if id != "" {
-				sessions[id] = struct{}{}
-			}
-		}
-		return nil
-	}
-
-	if err := collect(&models.ChatMessage{}); err != nil {
-		return nil, err
-	}
-	if err := collect(&models.UserWeightScore{}); err != nil {
-		return nil, err
-	}
-	if err := collect(&models.ConversationContext{}); err != nil {
-		return nil, err
-	}
-	if err := collect(&models.AIGeneratedQuestion{}); err != nil {
-		return nil, err
-	}
-	if err := collect(&models.UserAnalysisProgress{}); err != nil {
-		return nil, err
-	}
-	if err := collect(&models.UserEmbedding{}); err != nil {
-		return nil, err
-	}
-	if err := collect(&models.UserCompanyMatch{}); err != nil {
-		return nil, err
-	}
-	if err := collect(&models.VariantAssignment{}); err != nil {
-		return nil, err
-	}
-	if err := collect(&models.ResumeDocument{}); err != nil {
-		return nil, err
-	}
-
-	result := make([]string, 0, len(sessions))
-	for id := range sessions {
-		result = append(result, id)
-	}
-	return result, nil
-}
-
-func collectInterviewSessionIDs(tx *gorm.DB, userID uint) ([]uint, error) {
-	var sessionIDs []uint
-	if err := tx.Model(&models.InterviewSession{}).
-		Where("user_id = ?", userID).
-		Pluck("id", &sessionIDs).Error; err != nil {
-		return nil, err
-	}
-	return sessionIDs, nil
-}
-
-func collectResumeDocumentIDs(tx *gorm.DB, userID uint) ([]uint, error) {
-	var documentIDs []uint
-	if err := tx.Model(&models.ResumeDocument{}).
-		Where("user_id = ?", userID).
-		Pluck("id", &documentIDs).Error; err != nil {
-		return nil, err
-	}
-	return documentIDs, nil
-}
-
-func collectResumeReviewIDs(tx *gorm.DB, documentIDs []uint) ([]uint, error) {
-	var reviewIDs []uint
-	if err := tx.Model(&models.ResumeReview{}).
-		Where("document_id IN ?", documentIDs).
-		Pluck("id", &reviewIDs).Error; err != nil {
-		return nil, err
-	}
-	return reviewIDs, nil
-}
-
-// RegisterRequest ユーザー登録リクエスト
 type RegisterRequest struct {
 	Email                    string `json:"email"    validate:"required,email"`
 	Password                 string `json:"password" validate:"required,min=8"`
@@ -223,501 +115,4 @@ type AuthResponse struct {
 	RefreshToken             string `json:"refresh_token,omitempty"`  // リフレッシュトークン (#616)
 	EmailVerified            bool   `json:"email_verified"`
 	RequiresReVerification   bool   `json:"requires_re_verification,omitempty"`
-}
-
-// RequestRegistration メールアドレスに確認URLを送信して仮登録を作成
-func (s *AuthService) RequestRegistration(email string) error {
-	if email == "" {
-		return errors.New("email is required")
-	}
-
-	// 既存ユーザーチェック
-	existing, err := s.userRepo.GetUserByEmail(email)
-	if err != nil {
-		return fmt.Errorf("failed to check existing user: %w", err)
-	}
-	if existing != nil {
-		return errors.New("email already exists")
-	}
-
-	// 以前の仮登録を削除
-	if err := s.pendingRepo.DeleteByEmail(email); err != nil {
-		log.Printf("[AuthService] failed to delete previous pending registration for %s: %v", email, err)
-	}
-
-	// トークン生成
-	b := make([]byte, 32)
-	if _, err := rand.Read(b); err != nil {
-		return fmt.Errorf("failed to generate token: %w", err)
-	}
-	token := base64.URLEncoding.EncodeToString(b)
-
-	pending := &entity.PendingRegistration{
-		Token:     token,
-		Email:     email,
-		ExpiresAt: time.Now().Add(config.PendingRegistrationTokenTTL),
-	}
-	if err := s.pendingRepo.Create(pending); err != nil {
-		return fmt.Errorf("failed to save pending registration: %w", err)
-	}
-
-	return s.emailService.SendRegistrationEmail(email, token)
-}
-
-// ValidateRegistrationToken 仮登録トークンを検証してメールアドレスを返す
-func (s *AuthService) ValidateRegistrationToken(token string) (string, error) {
-	pending, err := s.pendingRepo.FindByToken(token)
-	if err != nil {
-		return "", fmt.Errorf("failed to find token: %w", err)
-	}
-	if pending == nil {
-		return "", errors.New("invalid or expired token")
-	}
-	return pending.Email, nil
-}
-
-// Register 新規ユーザー登録
-func (s *AuthService) Register(req RegisterRequest) (*AuthResponse, error) {
-	// バリデーション
-	if req.Email == "" || req.Password == "" {
-		return nil, errors.New("email and password are required")
-	}
-	if len(req.Password) < 8 {
-		return nil, errors.New("password must be at least 8 characters")
-	}
-
-	// トークン検証
-	if req.RegistrationToken != "" {
-		pending, err := s.pendingRepo.FindByToken(req.RegistrationToken)
-		if err != nil {
-			return nil, fmt.Errorf("failed to validate token: %w", err)
-		}
-		if pending == nil || pending.Email != req.Email {
-			return nil, errors.New("invalid or expired registration token")
-		}
-		// 使用済みトークンを削除
-		if err := s.pendingRepo.DeleteByEmail(req.Email); err != nil {
-			log.Printf("[AuthService] failed to delete used registration token for %s: %v", req.Email, err)
-		}
-	}
-	if req.TargetLevel == "" {
-		req.TargetLevel = "新卒"
-	}
-	if req.TargetLevel != "新卒" && req.TargetLevel != "中途" {
-		return nil, errors.New("target_level must be '新卒' or '中途'")
-	}
-	if strings.TrimSpace(req.SchoolName) == "" {
-		req.SchoolName = config.SchoolName()
-	}
-
-	// 既存ユーザーチェック
-	existingUser, err := s.userRepo.GetUserByEmail(req.Email)
-	if err != nil {
-		return nil, fmt.Errorf("failed to check existing user: %w", err)
-	}
-	if existingUser != nil {
-		return nil, errors.New("email already exists")
-	}
-
-	// パスワードハッシュ化
-	hashedPassword, err := bcrypt.GenerateFromPassword([]byte(req.Password), bcryptCost)
-	if err != nil {
-		return nil, fmt.Errorf("failed to hash password: %w", err)
-	}
-
-	// ユーザー作成
-	user := &entity.User{
-		Email:                    req.Email,
-		Password:                 string(hashedPassword),
-		Name:                     req.Name,
-		IsGuest:                  false,
-		TargetLevel:              req.TargetLevel,
-		SchoolName:               req.SchoolName,
-		IsAdmin:                  false,
-		CertificationsAcquired:   req.CertificationsAcquired,
-		CertificationsInProgress: req.CertificationsInProgress,
-	}
-
-	// メール認証トークン生成（有効期限 24 時間）（#330）
-	tokenBytes := make([]byte, 24)
-	if _, err := rand.Read(tokenBytes); err != nil {
-		return nil, fmt.Errorf("failed to generate verification token: %w", err)
-	}
-	user.EmailVerificationToken = base64.URLEncoding.EncodeToString(tokenBytes)
-	emailVerExpires := time.Now().Add(24 * time.Hour)
-	user.EmailVerificationExpires = &emailVerExpires
-
-	if err := s.userRepo.CreateUser(user); err != nil {
-		return nil, fmt.Errorf("failed to create user: %w", err)
-	}
-
-	// 認証メール送信（失敗しても登録は成功扱い）
-	appURL := config.AppURL()
-	go func() {
-		if err := s.emailService.SendVerificationEmail(user, user.EmailVerificationToken, appURL); err != nil {
-			log.Printf("[AuthService] failed to send verification email to %s: %v", user.Email, err)
-		}
-	}()
-
-	return &AuthResponse{
-		UserID:                   user.ID,
-		Email:                    user.Email,
-		Name:                     user.Name,
-		IsGuest:                  user.IsGuest,
-		TargetLevel:              user.TargetLevel,
-		SchoolName:               user.SchoolName,
-		IsAdmin:                  user.IsAdmin,
-		CertificationsAcquired:   user.CertificationsAcquired,
-		CertificationsInProgress: user.CertificationsInProgress,
-		EmailVerified:            false,
-	}, nil
-}
-
-// Login ログイン処理
-func (s *AuthService) Login(req LoginRequest) (*AuthResponse, error) {
-	// バリデーション
-	if req.Email == "" || req.Password == "" {
-		return nil, errors.New("email and password are required")
-	}
-
-	// ユーザー取得
-	user, err := s.userRepo.GetUserByEmail(req.Email)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get user: %w", err)
-	}
-	if user == nil {
-		return nil, errors.New("invalid email or password")
-	}
-	if user.IsWithdrawn() {
-		return nil, errors.New("account has been withdrawn")
-	}
-
-	// ゲストユーザーはログイン不可
-	if user.IsGuest {
-		return nil, errors.New("guest users cannot login")
-	}
-
-	// パスワード検証
-	if err := bcrypt.CompareHashAndPassword([]byte(user.Password), []byte(req.Password)); err != nil {
-		return nil, errors.New("invalid email or password")
-	}
-
-	// コストが古い場合は再ハッシュ（#328: bcryptCost=12 への移行）
-	if cost, err := bcrypt.Cost([]byte(user.Password)); err == nil && cost < bcryptCost {
-		if rehashed, err := bcrypt.GenerateFromPassword([]byte(req.Password), bcryptCost); err == nil {
-			user.Password = string(rehashed)
-			_ = s.userRepo.UpdateUser(user)
-		}
-	}
-
-	isOAuth := user.OAuthProvider != ""
-	emailVerified := user.EmailVerifiedAt != nil
-	requiresReVerification := false
-
-	if !isOAuth {
-		// メール認証チェック
-		if !emailVerified {
-			return nil, errors.New("email_not_verified")
-		}
-
-		// 10日以上ログインなし → 再認証
-		if user.LastLoginAt != nil && time.Since(*user.LastLoginAt) > config.ReVerificationInactiveDuration {
-			tokenBytes := make([]byte, 24)
-			if _, err := rand.Read(tokenBytes); err != nil {
-				return nil, fmt.Errorf("failed to generate re-verification token: %w", err)
-			}
-			user.EmailVerificationToken = base64.URLEncoding.EncodeToString(tokenBytes)
-			reVerExpires := time.Now().Add(24 * time.Hour)
-			user.EmailVerificationExpires = &reVerExpires
-			user.EmailVerifiedAt = nil
-			s.userRepo.UpdateUser(user)
-			appURL := config.AppURL()
-			go s.emailService.SendReVerificationEmail(user, user.EmailVerificationToken, appURL)
-			requiresReVerification = true
-			return nil, errors.New("re_verification_required")
-		}
-	}
-
-	// 最終ログイン更新
-	now := time.Now()
-	user.LastLoginAt = &now
-	err = s.userRepo.UpdateUser(user)
-	if err != nil {
-		return nil, err
-	}
-
-	resp := &AuthResponse{
-		UserID:                   user.ID,
-		Email:                    user.Email,
-		Name:                     user.Name,
-		IsGuest:                  user.IsGuest,
-		TargetLevel:              user.TargetLevel,
-		SchoolName:               user.SchoolName,
-		IsAdmin:                  user.IsAdmin,
-		CertificationsAcquired:   user.CertificationsAcquired,
-		CertificationsInProgress: user.CertificationsInProgress,
-		AvatarURL:                user.AvatarURL,
-		EmailVerified:            emailVerified,
-		RequiresReVerification:   requiresReVerification,
-	}
-	adminSecret := os.Getenv("ADMIN_SECRET")
-	userSecret := os.Getenv("USER_SECRET")
-	// 管理者ユーザーにはHMACトークンを付与する
-	if user.IsAdmin && adminSecret != "" {
-		resp.Token = middleware.GenerateAdminToken(user.ID, user.Email, adminSecret)
-	}
-	// 全ユーザーにユーザー認証トークンを付与する
-	if userSecret != "" {
-		resp.UserToken = middleware.GenerateUserToken(user.ID, user.Email, userSecret)
-		resp.RefreshToken = s.issueRefreshToken(user.ID)
-	}
-	return resp, nil
-}
-
-// CreateGuestUser ゲストユーザー作成
-func (s *AuthService) CreateGuestUser() (*AuthResponse, error) {
-	// ランダムなゲストID生成
-	randomBytes := make([]byte, 16)
-	if _, err := rand.Read(randomBytes); err != nil {
-		return nil, fmt.Errorf("failed to generate random ID: %w", err)
-	}
-	guestID := base64.URLEncoding.EncodeToString(randomBytes)
-
-	user := &entity.User{
-		Email:       fmt.Sprintf("guest_%s@%s", guestID, config.GuestEmailDomain()),
-		Password:    "", // ゲストユーザーはパスワード不要
-		Name:        fmt.Sprintf("Guest_%s", guestID[:8]),
-		IsGuest:     true,
-		TargetLevel: "未設定",
-		SchoolName:  config.SchoolName(),
-	}
-
-	if err := s.userRepo.CreateUser(user); err != nil {
-		return nil, fmt.Errorf("failed to create guest user: %w", err)
-	}
-
-	resp := &AuthResponse{
-		UserID:                   user.ID,
-		Email:                    user.Email,
-		Name:                     user.Name,
-		IsGuest:                  user.IsGuest,
-		TargetLevel:              user.TargetLevel,
-		SchoolName:               user.SchoolName,
-		IsAdmin:                  user.IsAdmin,
-		CertificationsAcquired:   user.CertificationsAcquired,
-		CertificationsInProgress: user.CertificationsInProgress,
-		AvatarURL:                user.AvatarURL,
-	}
-	userSecret := os.Getenv("USER_SECRET")
-	if userSecret != "" {
-		resp.UserToken = middleware.GenerateUserToken(user.ID, user.Email, userSecret)
-		resp.RefreshToken = s.issueRefreshToken(user.ID)
-	}
-	return resp, nil
-}
-
-// RefreshSession はリフレッシュトークンをローテーションし、新しいトークンペアを返す (#616)
-func (s *AuthService) RefreshSession(refreshToken string) (*AuthResponse, error) {
-	if s.refreshTokens == nil {
-		return nil, errors.New("refresh token service not configured")
-	}
-	userSecret := os.Getenv("USER_SECRET")
-	if userSecret == "" {
-		return nil, errors.New("USER_SECRET is not configured")
-	}
-
-	userID, newRefreshToken, err := s.refreshTokens.Rotate(refreshToken)
-	if err != nil {
-		return nil, err
-	}
-
-	user, err := s.userRepo.GetUserByID(userID)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get user: %w", err)
-	}
-	// 退会済みユーザーはリフレッシュ不可 (#613)
-	if user == nil || user.IsWithdrawn() {
-		return nil, ErrInvalidRefreshToken
-	}
-
-	return &AuthResponse{
-		UserID:       user.ID,
-		Email:        user.Email,
-		Name:         user.Name,
-		IsGuest:      user.IsGuest,
-		IsAdmin:      user.IsAdmin,
-		UserToken:    middleware.GenerateUserToken(user.ID, user.Email, userSecret),
-		RefreshToken: newRefreshToken,
-	}, nil
-}
-
-// LogoutSession はリフレッシュトークンを失効させる (#616)
-func (s *AuthService) LogoutSession(refreshToken string) error {
-	if s.refreshTokens == nil || refreshToken == "" {
-		return nil
-	}
-	return s.refreshTokens.Revoke(refreshToken)
-}
-
-// GetUser ユーザー情報取得
-func (s *AuthService) GetUser(userID uint) (*AuthResponse, error) {
-	user, err := s.userRepo.GetUserByID(userID)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get user: %w", err)
-	}
-	if user == nil {
-		return nil, errors.New("user not found")
-	}
-
-	return &AuthResponse{
-		UserID:                   user.ID,
-		Email:                    user.Email,
-		Name:                     user.Name,
-		IsGuest:                  user.IsGuest,
-		TargetLevel:              user.TargetLevel,
-		SchoolName:               user.SchoolName,
-		IsAdmin:                  user.IsAdmin,
-		CertificationsAcquired:   user.CertificationsAcquired,
-		CertificationsInProgress: user.CertificationsInProgress,
-		AvatarURL:                user.AvatarURL,
-		OAuthProvider:            user.OAuthProvider,
-	}, nil
-}
-
-// UpdateProfile ユーザープロフィール更新
-func (s *AuthService) UpdateProfile(req UpdateProfileRequest) (*AuthResponse, error) {
-	if req.UserID == 0 {
-		return nil, errors.New("user_id is required")
-	}
-	if req.TargetLevel != "" && req.TargetLevel != "新卒" && req.TargetLevel != "中途" {
-		return nil, errors.New("target_level must be '新卒' or '中途'")
-	}
-
-	user, err := s.userRepo.GetUserByID(req.UserID)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get user: %w", err)
-	}
-	if user == nil {
-		return nil, errors.New("user not found")
-	}
-
-	if req.Name != "" {
-		user.Name = req.Name
-	}
-	if req.TargetLevel != "" {
-		user.TargetLevel = req.TargetLevel
-	}
-	// Always persist the provided school name, even when it is an empty string.
-	user.SchoolName = req.SchoolName
-	user.CertificationsAcquired = req.CertificationsAcquired
-	user.CertificationsInProgress = req.CertificationsInProgress
-
-	if err := s.userRepo.UpdateUser(user); err != nil {
-		return nil, fmt.Errorf("failed to update user: %w", err)
-	}
-
-	return &AuthResponse{
-		UserID:                   user.ID,
-		Email:                    user.Email,
-		Name:                     user.Name,
-		IsGuest:                  user.IsGuest,
-		TargetLevel:              user.TargetLevel,
-		SchoolName:               user.SchoolName,
-		IsAdmin:                  user.IsAdmin,
-		CertificationsAcquired:   user.CertificationsAcquired,
-		CertificationsInProgress: user.CertificationsInProgress,
-		AvatarURL:                user.AvatarURL,
-	}, nil
-}
-
-// RequestPasswordReset パスワードリセットメールを送信
-func (s *AuthService) RequestPasswordReset(email string) error {
-	user, err := s.userRepo.GetUserByEmail(email)
-	if err != nil {
-		return fmt.Errorf("failed to get user: %w", err)
-	}
-	// ユーザーが存在しない・OAuthユーザー・ゲストの場合でも成功を返す（情報漏洩防止）
-	if user == nil || user.OAuthProvider != "" || user.IsGuest {
-		return nil
-	}
-
-	// 32バイトのランダムトークンを生成
-	b := make([]byte, 32)
-	if _, err := rand.Read(b); err != nil {
-		return fmt.Errorf("failed to generate token: %w", err)
-	}
-	token := base64.URLEncoding.EncodeToString(b)
-
-	expiresAt := time.Now().Add(config.PasswordResetTokenTTL)
-	user.PasswordResetToken = token
-	user.PasswordResetExpiresAt = &expiresAt
-
-	if err := s.userRepo.UpdateUser(user); err != nil {
-		return fmt.Errorf("failed to update user: %w", err)
-	}
-
-	appURL := config.AppURL()
-	return s.emailService.SendPasswordResetEmail(user.Email, token, appURL)
-}
-
-// ResetPassword トークンを検証して新パスワードをセット
-func (s *AuthService) ResetPassword(token, newPassword string) error {
-	if token == "" {
-		return errors.New("token is required")
-	}
-
-	user, err := s.userRepo.GetUserByPasswordResetToken(token)
-	if err != nil {
-		return fmt.Errorf("failed to find user: %w", err)
-	}
-	if user == nil {
-		return errors.New("invalid or expired token")
-	}
-	if user.PasswordResetExpiresAt == nil || time.Now().After(*user.PasswordResetExpiresAt) {
-		return errors.New("invalid or expired token")
-	}
-
-	hashedPassword, err := bcrypt.GenerateFromPassword([]byte(newPassword), bcryptCost)
-	if err != nil {
-		return fmt.Errorf("failed to hash password: %w", err)
-	}
-
-	user.Password = string(hashedPassword)
-	user.PasswordResetToken = ""
-	user.PasswordResetExpiresAt = nil
-
-	if err := s.userRepo.UpdateUser(user); err != nil {
-		return err
-	}
-
-	// パスワード変更時は全端末のリフレッシュトークンを失効させる (#616)
-	if s.refreshTokens != nil {
-		if err := s.refreshTokens.RevokeAllForUser(user.ID); err != nil {
-			log.Printf("refresh token revoke all failed user_id=%d error=%v", user.ID, err)
-		}
-	}
-	return nil
-}
-
-// VerifyEmail トークンを検証してメールを認証済みにする（#330: 有効期限チェック追加）
-func (s *AuthService) VerifyEmail(token string) error {
-	if token == "" {
-		return errors.New("token is required")
-	}
-	user, err := s.userRepo.GetUserByVerificationToken(token)
-	if err != nil {
-		return fmt.Errorf("failed to find user: %w", err)
-	}
-	if user == nil {
-		return errors.New("invalid or expired token")
-	}
-	// 有効期限チェック（ExpiresAt が設定されている場合のみ）
-	if user.EmailVerificationExpires != nil && time.Now().After(*user.EmailVerificationExpires) {
-		return errors.New("invalid or expired token")
-	}
-	now := time.Now()
-	user.EmailVerifiedAt = &now
-	user.EmailVerificationToken = ""
-	user.EmailVerificationExpires = nil
-	return s.userRepo.UpdateUser(user)
 }
