+++ Backend/internal/services/auth_service_test.go
@@ -86,3 +86,50 @@ func TestDeleteAccount_AlreadyWithdrawn(t *testing.T) {
 		t.Fatalf("unmet sqlmock expectations: %v", err)
 	}
 }
+
+func TestCollectInterviewAndResumeIDs(t *testing.T) {
+	db, mock := newAuthServiceTestDB(t)
+
+	mock.ExpectQuery("SELECT `id` FROM `interview_sessions` WHERE user_id = \\?").
+		WithArgs(uint(9)).
+		WillReturnRows(sqlmock.NewRows([]string{"id"}).AddRow(3).AddRow(5))
+	mock.ExpectQuery("SELECT `id` FROM `resume_documents` WHERE user_id = \\?").
+		WithArgs(uint(9)).
+		WillReturnRows(sqlmock.NewRows([]string{"id"}).AddRow(11))
+	mock.ExpectQuery("SELECT `id` FROM `resume_reviews` WHERE document_id IN \\(\\?\\)").
+		WithArgs(uint(11)).
+		WillReturnRows(sqlmock.NewRows([]string{"id"}).AddRow(21))
+
+	interviewIDs, err := collectInterviewSessionIDs(db, 9)
+	if err != nil {
+		t.Fatalf("collectInterviewSessionIDs: %v", err)
+	}
+	if len(interviewIDs) != 2 || interviewIDs[0] != 3 || interviewIDs[1] != 5 {
+		t.Fatalf("interviewIDs=%v", interviewIDs)
+	}
+
+	docIDs, err := collectResumeDocumentIDs(db, 9)
+	if err != nil {
+		t.Fatalf("collectResumeDocumentIDs: %v", err)
+	}
+	if len(docIDs) != 1 || docIDs[0] != 11 {
+		t.Fatalf("docIDs=%v", docIDs)
+	}
+
+	reviewIDs, err := collectResumeReviewIDs(db, docIDs)
+	if err != nil {
+		t.Fatalf("collectResumeReviewIDs: %v", err)
+	}
+	if len(reviewIDs) != 1 || reviewIDs[0] != 21 {
+		t.Fatalf("reviewIDs=%v", reviewIDs)
+	}
+	if err := mock.ExpectationsWereMet(); err != nil {
+		t.Fatalf("unmet: %v", err)
+	}
+}
+
+func TestBcryptCostIsOWASPRecommended(t *testing.T) {
+	if bcryptCost < 12 {
+		t.Fatalf("bcryptCost=%d want >= 12", bcryptCost)
+	}
+}
